+++ R/globals.R
@@ -7,5 +7,5 @@ utils::globalVariables(c(
   "label_to", "missing_from", "missing_to", "rank", "rank_change",
   "movement", "rank_from", "rank_to", "rank_x", "status", "to", "transition_index",
   "value_change", "value_from",
-  "value_left", "value_right", "value_to", "value_x", "x", "y", "y_next"
+  "link_right", "value_left", "value_right", "value_to", "value_x", "x", "y", "y_next"
 ))